Rename stimulus controller.
