+++ app/javascript/spree/dashboard/controllers/upload_button_controller.js
